requirements.txt
absl-py==0.12.0
alembic==1.7.6
anyio==3.3.4
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
bokeh==2.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
cryptography==36.0.1
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
dm-tree==0.1.6
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.73.0
filelock==3.4.2
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
gym==0.21.0
h11==0.13.0
h5py==3.5.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.3
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
orjson==3.6.6
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.9.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.2
starlette==0.17.1
sympy==1.9
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
tensorflow-probability==0.14.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tfp-nightly==0.15.0.dev20211115
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.13.0
voila==0.2.16
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.3
xmltodict==0.12.0
